feat: explode code across helpers and classes
